requirements.txt
gunicorn
django-heroku
requests
django-allauth
django-session-timeout
django-materializecss-form
black
coverage
coveralls
flake8